Extract helper functions to enable testing actual code
Move pure helper functions from view.js modules into separate helpers.js
files that can be imported without Interactivity API dependencies. This
allows Jest tests to import and test the actual production code instead
of maintaining duplicated copies of the functions.
