+++ projects/packages/forms/src/modules/form/view.js
@@ -12,7 +12,8 @@ import {
  * Internal dependencies
  */
 import { validateField, isEmptyValue } from '../../contact-form/js/validate-helper.js';
-import { getRating, getRatingDisplayValue } from '../field-rating/view.js';
+import { getRating } from '../field-rating/view.js';
+import { maybeAddColonToLabel, maybeTransformValue, getImages, getUrl } from './helpers.js';
 import { focusNextInput, submitForm } from './shared.ts';
 
 const withSyncEvent =
@@ -114,86 +115,6 @@ const getError = field => {
 	return config.error_types && config.error_types[ field.error ];
 };
 
-const maybeAddColonToLabel = label => {
-	const formattedLabel = label ? label : null;
-
-	if ( ! formattedLabel ) {
-		return null;
-	}
-	// Special case for the Terms consent field block which has a period at the end of the text.
-	return formattedLabel.endsWith( '?' )
-		? formattedLabel
-		: formattedLabel.replace( /[.:]$/, '' ) + ':';
-};
-
-const maybeTransformValue = value => {
-	// For image select fields, we want to show the perceived values, as the choices can be shuffled.
-	if ( value?.type === 'image-select' ) {
-		return value.choices
-			.map( choice => {
-				let transformedValue = choice.perceived;
-
-				if ( choice.showLabels && choice.label != null && choice.label !== '' ) {
-					transformedValue += ' - ' + choice.label;
-				}
-
-				return transformedValue;
-			} )
-			.join( ', ' );
-	}
-
-	// For URL fields, extract the URL text value.
-	if ( value?.type === 'url' && value?.url ) {
-		return value.url;
-	}
-
-	// For rating fields, return the displayValue (e.g., "3/5") for text fallback.
-	const ratingDisplayValue = getRatingDisplayValue( value );
-	if ( ratingDisplayValue ) {
-		return ratingDisplayValue;
-	}
-
-	// For file upload fields, we want to show the file name and size
-	if ( value?.name && value?.size ) {
-		return value.name + ' (' + value.size + ')';
-	}
-
-	return value;
-};
-
-const getImages = value => {
-	if ( value?.type === 'image-select' ) {
-		return value.choices.map( choice => {
-			const letterCode = choice.perceived ?? '';
-			const label =
-				choice.showLabels && choice.label != null && choice.label !== '' ? choice.label : '';
-
-			return {
-				src: choice.image?.src ?? '',
-				letterCode,
-				label,
-			};
-		} );
-	}
-
-	return null;
-};
-
-const getUrl = value => {
-	if ( value?.type === 'url' && value?.url ) {
-		let url = value.url;
-
-		// Prepend https:// if no protocol is specified.
-		if ( ! /^https?:\/\//i.test( url ) ) {
-			url = 'https://' + url;
-		}
-
-		return url;
-	}
-
-	return null;
-};
-
 /**
  * Capture file preview data (thumbnail URLs and icons) from the DOM before form submission.
  * This allows us to preserve the client-side preview for the confirmation page.
@@ -270,6 +191,7 @@ const getFiles = value => {
 
 	return null;
 };
+
 
 const toggleImageOptionInput = ( input, optionElement ) => {
 	if ( input ) {
